removed unnecessary order form view; consolidated form logic in order list view
