+++ src/views/order_list_view.js
@@ -1,14 +1,16 @@
 import Backbone from 'backbone';
 
+import Order from '../models/order';
 import OrderView from './order_view';
 
 const OrderListView = Backbone.View.extend({
   initialize(params) {
     this.template = params.template;
     this.bus = params.bus;
+    this.quotes = params.quotes;
 
     this.listenTo(this.model, 'update', this.render);
-    this.listenTo(this.bus, 'addOrder', this.addOrder);
+    // this.listenTo(this.bus, 'addOrder', this.addOrder);
   },
   render() {
     this.$('#orders').empty();
@@ -26,6 +28,70 @@ const OrderListView = Backbone.View.extend({
     });
     return this;
   },
+  events: {
+    'click .btn-buy': 'buyOrder',
+    'click .btn-sell': 'sellOrder',
+  },
+
+  createOrder(buyOption) {
+    // const sym = this.$('#symbol').val();
+    // console.log(`sym is ${sym}`);
+    const orderData = {
+      quotes: this.quotes,
+      bus: this.bus,
+      symbol: this.$('#symbol').val(),
+      targetPrice: parseFloat(this.$('#target-price').val()),
+      // targetPrice: parseFloat(this.$('form input[name="price-target"]').val()),
+      buy: buyOption.buy,
+      // marketPrice: this.getCurrentPrice(this.getQuote(sym))
+    };
+
+    return new Order(orderData);
+  },
+
+  buyOrder: function(event) {
+    event.preventDefault();
+    this.clearErrors();
+    const order = this.createOrder({buy: true});
+
+    this.validate(order);
+    // this.$el.find('form').trigger('reset');
+  },
+
+  sellOrder: function(event) {
+    event.preventDefault();
+    this.clearErrors();
+
+    const order = this.createOrder({buy: false});
+
+    this.validate(order);
+    // this.$el.find('form').trigger('reset');
+  },
+
+  validate(order) {
+    if (order.isValid()) {
+      this.model.add(order);
+      this.$el.find('form').trigger('reset');
+      // this.bus.trigger('addOrder', order);
+    } else {
+      this.renderError(order.validationError);
+    }
+  },
+
+  clearErrors() {
+    this.$('.form-errors').empty();
+  },
+
+  renderError(errors) {
+    const errorSection = this.$('.form-errors');
+    Object.keys(errors).forEach((field) => {
+      errors[field].forEach((error) => {
+        const html = `<p class="error-message small-6 cell">${field}: ${error}</p>`;
+        errorSection.append(html);
+      });
+    });
+  },
+
   addOrder(order) {
     this.model.add(order);
   },
